spiritual-voice-bot/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & AI
anthropic==0.39.0
transformers==4.36.2
torch==2.1.2
accelerate==0.26.1
bitsandbytes==0.41.3
sentence-transformers==2.3.1
langchain==0.1.4
langchain-community>=0.0.14

# RAG & Vector DB
qdrant-client==1.7.0
chromadb==0.4.22
faiss-cpu==1.7.4

# Voice Processing
openai-whisper==20231117
TTS==0.22.0
gtts==2.5.1
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1

# Indic NLP
indic-nlp-library==0.92
ai4bharat-transliteration==1.0.0

# Data Processing
pandas==1.5.3
numpy==1.22.0
datasets==2.16.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.26.0
pyyaml==6.0.1
redis==5.0.1

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
